+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `!=` is correctly recognized by the lexer
+
 ## 0.3.1 - 2025-11-21
 
 ### Added